@@ -1,54 +1,43 @@
 {
-  "name": "omc",
+  "name": "o-media-center.web",
   "version": "0.0.0",
-  "license": "MIT",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build --prod",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.4",
-    "@angular/cdk": "^6.2.1",
-    "@angular/common": "6.0.4",
-    "@angular/compiler": "^6.0.4",
-    "@angular/core": "^6.0.4",
-    "@angular/forms": "^6.0.4",
-    "@angular/http": "^6.0.4",
-    "@angular/material": "^6.2.1",
-    "@angular/platform-browser": "^6.0.4",
-    "@angular/platform-browser-dynamic": "^6.0.4",
-    "@angular/router": "^6.0.4",
-    "core-js": "^2.4.1",
-    "hls.js": "^0.9.1",
-    "rxjs": "^6.2.1",
-    "videogular2": "^6.2.2",
-    "zone.js": "^0.8.19"
+    "@angular/animations": "~13.1.0",
+    "@angular/cdk": "^13.1.2",
+    "@angular/common": "~13.1.0",
+    "@angular/compiler": "~13.1.0",
+    "@angular/core": "~13.1.0",
+    "@angular/forms": "~13.1.0",
+    "@angular/material": "^13.1.2",
+    "@angular/platform-browser": "~13.1.0",
+    "@angular/platform-browser-dynamic": "~13.1.0",
+    "@angular/router": "~13.1.0",
+    "@videogular/ngx-videogular": "^5.0.1",
+    "hls.js": "^1.1.2",
+    "rxjs": "~7.4.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "^6.0.8",
-    "@angular/compiler-cli": "^6.0.4",
-    "@angular/language-service": "^6.0.4",
-    "@types/core-js": "^2.5.0",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "@angular-devkit/build-angular": "~13.1.2",
+    "@angular/cli": "~13.1.2",
+    "@angular/compiler-cli": "~13.1.0",
+    "@types/jasmine": "~3.10.0",
+    "@types/node": "^12.11.1",
+    "jasmine-core": "~3.10.0",
+    "karma": "~6.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.1.0",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "typescript": "~4.5.2"
   }
 }
