@@ -1,5 +1,5 @@
 export interface ClientCommand {
     command: string;
-    parameter: string;
+    parameter: string | null;
     date: Date;
 }
