@@ -1,9 +1,14 @@
 export interface MediaFileRecord {
-    name: string;
-    description: string;
-    hash: string;
-    mediaType: string;
-    thumbnailType: string;
-    technicalInfo: string;
-    lastPlayedTime: number;
+  name: string;
+  description: string;
+  hash: string;
+  mediaType: string;
+  thumbnailType: string;
+  technicalInfo: string;
+  lastPlayedTime: number;
+  year: string;
+  season: string;
+  episode: string;
+  imdbNumber: string;
+  genres: string;
 }
