@@ -1,33 +1,36 @@
 import { Injectable } from '@angular/core';
-import { Http, Headers } from '@angular/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 
 @Injectable()
 export class SettingsService {
 
-  private headers = new Headers({ 'Content-Type': 'application/json' });
+  private headers = new HttpHeaders({ 'Content-Type': 'application/json' });
   private clientCommandUrl = 'api/v1/client';  // URL to web api
 
-  private clientId: string;
+  private clientId!: string | null;
 
-  constructor(private http: Http) {
+  constructor(private http: HttpClient) {
   }
 
-  getClientId() : Promise<string> {
-  if (localStorage.getItem('clientId')) {
-          this.clientId = localStorage.getItem('clientId');
-          return new Promise<string>(r => r(this.clientId));
-      } else {
-          return this.http.post(this.clientCommandUrl, { ClientName: '' })
-              .toPromise().then(response => {
-                  var clientId = response.text();
-                  localStorage.setItem('clientId', clientId);
-                  return clientId;
-              });
-      }
+  getClientId(): Promise<string | null> {
+    if (localStorage.getItem('clientId')) {
+      this.clientId = localStorage.getItem('clientId');
+      return new Promise<string | null>(r => r(this.clientId));
+    } else {
+      return this.http.post(this.clientCommandUrl, { ClientName: '' })
+        .toPromise().then(response => {
+          if (response) {
+            var clientId = response as string;
+            localStorage.setItem('clientId', clientId);
+            return clientId;
+          }
+          return null;
+        });
+    }
   }
 
   setClientId(clientId: string) {
-      localStorage.setItem('clientId', clientId);
+    localStorage.setItem('clientId', clientId);
   }
 
 }
